django/requirements.txt
Django==3.2.9
gunicorn==20.1.0
psycopg2-binary==2.9.2
drf-writable-nested==0.6.3
djangorestframework==3.12.4
django-cors-headers==3.10.0
django-storages==1.12.3
djangorestframework-simplejwt==5.2.2